+++ R/timemachine.R
@@ -5,24 +5,35 @@
 #' @param from Date new version of package. Default: NULL
 #' @param version character version of package. Default: NULL
 #' @param to Date CRAN URL. Default: NULL
-#' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
-#' @return data.frame
+#' @param repos character the base URL of the repository to use. Used only for the validation. Default `https://cran.rstudio.com/`
+#' @return data.frame with 7 columns
+#' \describe{
+#' \item{Package}{character package name.}
+#' \item{Version}{character package version.}
+#' \item{Released}{character release Date}
+#' \item{Archived}{character archived Date.}
+#' \item{LifeDuration}{`difftime` number of days the version was the newest one.}
+#' \item{URL}{character the suffix of the base URL to tar.gz file.}
+#' \item{Size}{character size of the `tar.gz` file.}
+#' }
 #' @note Function will scrap two CRAN URLS. Works only with CRAN packages.
 #' Please as a courtesy to the R CRAN, don't overload their servers by constantly using this function.
 #' The base part of URL in the result is `https://cran.r-project.org/src/contrib/`.
 #' Results are cached for 1 hour with `memoise` package.
 #' @export
 #' @examples
+#' \dontrun{
 #' pac_timemachine("dplyr", at = as.Date("2017-02-02"))
 #' pac_timemachine("dplyr", from = as.Date("2017-02-02"), to = as.Date("2018-04-02"))
 #' pac_timemachine("dplyr", at = Sys.Date())
+#' }
 pac_timemachine <- function(pac,
                             at = NULL,
                             from = NULL,
                             to = NULL,
                             version = NULL,
                             repos = "https://cran.rstudio.com/") {
-  stopifnot(pac %in% c(rownames(available_packages(repos = repos)), pacs_base()))
+  stopifnot(pac %in% rownames(available_packages(repos = repos)))
   stopifnot(is.null(version) || (length(version) == 1 && is.character(version)))
   stopifnot(xor(
     !is.null(at) && inherits(at, "Date") && is.null(version),
@@ -38,24 +49,24 @@ pac_timemachine <- function(pac,
   }
 
   result$Archived <- as.Date(c(result$Released[-1], cran_page$Released))
-  result$Life_Duration <- result$Archived - result$Released
-  f_cols <- c("Package", "Version", "Released", "Archived", "Life_Duration", "URL", "Size")
+  result$LifeDuration <- result$Archived - result$Released
+  f_cols <- c("Package", "Version", "Released", "Archived", "LifeDuration", "URL", "Size")
   result <- rbind(result[, f_cols], cran_page[, f_cols])
   result <- result[, f_cols]
 
-  if (!is.null(at)) {
-    if (all(at >= result$Released)) {
+  if (isTRUE(!is.null(at))) {
+    if (isTRUE(all(at >= result$Released))) {
       utils::tail(result, 1)
     } else {
       result[at >= result$Released & at <= result$Archived, ]
     }
-  } else if (!is.null(from) && !is.null(to)) {
+  } else if (isTRUE(!is.null(from) && !is.null(to))) {
     if (all(from >= result$Released)) {
       utils::tail(result, 1)
     } else {
       result[to >= result$Released & from <= result$Archived, ]
     }
-  } else if (!is.null(version)) {
+  } else if (isTRUE(!is.null(version))) {
     result[result$Version == version, ]
   } else {
     result
@@ -64,31 +75,33 @@ pac_timemachine <- function(pac,
 
 pac_cran_recent_raw <- function(pac) {
   cran_page <- try(readLines(sprintf("https://CRAN.R-project.org/package=%s", pac)), silent = TRUE)
-  if (!inherits(cran_page, "try-error")) {
-    cran_v <- gsub("</?td>", "", cran_page[grep("Version:", cran_page) + 1])
-    cran_released <- gsub("</?td>", "", cran_page[grep("Published:", cran_page) + 1])
+  if (!inherits(cran_page, "try-error") && any(grepl(pac, cran_page))) {
+    cran_v <- utils::head(gsub("</?td>", "", cran_page[grep("Version:", cran_page) + 1]), 1)
+    cran_released <- utils::head(gsub("</?td>", "", cran_page[grep("Published:", cran_page) + 1]), 1)
+
+    if (length(cran_v) == 0) cran_v <- NA
+    if (length(cran_released) == 0) cran_released <- NA
+    f_cols <- c("Package", "Version", "Released", "Archived", "LifeDuration", "URL", "Size")
 
     data.frame(
       Package = pac,
-      Released = as.Date(cran_released),
-      Size = NA,
-      Description = NA,
       Version = cran_v,
+      Released = as.Date(cran_released),
       Archived = NA,
-      Life_Duration = Sys.Date() - as.Date(cran_released),
+      LifeDuration = Sys.Date() - as.Date(cran_released),
       URL = sprintf("%s_%s.tar.gz", pac, cran_v),
+      Size = NA,
       stringsAsFactors = FALSE
     )
   } else {
     data.frame(
       Package = pac,
-      Released = NA,
-      Size = NA,
-      Description = NA,
       Version = NA,
+      Released = NA,
       Archived = NA,
-      Life_Duration = NA,
+      LifeDuration = NA,
       URL = NA,
+      Size = NA,
       stringsAsFactors = FALSE
     )
   }
@@ -100,23 +113,17 @@ pac_archived_raw <- function(pac) {
   base_archive <- sprintf("/src/contrib/Archive/%s/", pac)
   rr <- try(readLines(paste0("https://cran.r-project.org", base_archive)), silent = TRUE)
 
-  if (!inherits(rr, "try-error")) {
+  if (!inherits(rr, "try-error") && any(grepl(pac, rr))) {
     rr_range <- grep("</?table>", rr)
     rrr <- rr[(rr_range[1] + 1):(rr_range[2] - 1)]
     # not use rvest as it is too big dependency
-    header <- stringi::stri_match_all(rrr[[1]], regex = ">([^<>]+)<")[[1]][, 2]
+    header <- trimws(xml_text(xml_find_all(read_html(rrr[1]), "//th")))
 
-    result_raw <- as.data.frame(
-      do.call(
-        rbind,
-        lapply(
-          4:(length(rrr) - 1),
-          function(x) stringi::stri_match_all(rrr[x], regex = ">([^<>]+)<")[[1]][, 2]
-        )
-      ),
-      stringsAsFactors = FALSE
-    )
-    colnames(result_raw) <- header
+    result_raw <- as.data.frame(matrix(trimws(xml_text(xml_find_all(read_html(paste(rrr[2:length(rrr)], collapse = "\n")), "//td"))),
+                         ncol = length(header),
+                         nrow = length(rrr) - 3, byrow = TRUE))
+    result_raw <- result_raw[-1, -1]
+    colnames(result_raw) <- header[-1]
 
     result <- result_raw[result_raw[["Last modified"]] != "", ]
     colnames(result) <- c("Package", "Released", "Size", "Description")